pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carbonscope"
version = "0.20.0"
description = "Bittensor subnet for decentralized corporate carbon emission estimation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "bittensor==10.1.0",
    "pydantic[email]>=2.0",
    "numpy>=1.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    "passlib[bcrypt]>=1.7.4",
    "PyJWT>=2.8.0",
    "httpx>=0.24.0",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "aiosmtplib>=3.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:asyncio",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
